Remove Basic.compare and 'old' print ordering
Also removes everything that depends on them. I haven't yet fixed all test
failures caused by this, or the previous commit.

The old order is heavily tied to Basic.compare and the ordering of classes. I
don't see a value in keeping it with a deprecation, as it would require also
keeping the ordering of classes. The name implies that it is de facto
deprecated, and it has been called 'old' for some time now. The main use-case
was supposedly that 'old' was faster, but I'm not even clear that that should
be the case (it still does sorting), and even if it is, if you care about
speed, you should use order='none'.

I don't think Basic.compare itself should be kept deprecated either. All uses
of it should use default_sort_key. The ordering will not be equivalent (which
is one reason why I don't think a deprecation would be that helpful), but the
ordering here is arbitrary anyway.